[GridNG] Fix failing scrollable overflow tests.
There were never passing in legacy, but is relatively trivial to pass
in NG.

Effectively NG has the concept of the "inflow-bounds" which for grid
layout is the size of the grid. Legacy never has this clear separation
(kinda using the intrinsic-block-size as a proxy).

Additionally fixes the tests (these have never passed in any browser
so had some "obvious" mistakes).

Bug: 1045599
Change-Id: I990f450ee7ec00a2fd730d402e25632b8c748cd3